fix(vald): broadcast messages sequentially and retry on failure (#344)
Concurrent broadcasts lead to failures because the tx sequence number would be set incorrectly. This commit makes the broadcaster more robust by sequencing all broadcast commands and introducing retry with exponential backoff when a message fails

(cherry picked from commit 3bf39621168cf9c9c6acffd42552ef26751b5150)

+++ cmd/vald/main.go
@@ -119,7 +119,7 @@ func loadConfig() (app.Config, string) {
 		panic(err)
 	}
 	// for some reason gas is not being filled
-	conf.Gas = viper.GetInt("gas")
+	conf.Gas = viper.GetUint64("gas")
 
 	return conf, viper.GetString("validator-addr")
 }
